Create 2.	Write a program to implement doubly linked list as an ADT that supports the following operations:
i.	Insert an element x at the beginning of the doubly linked list 
ii.	Insert an element x at the end of the doubly linked list 
iii.	Remove an element from the beginning of the doubly linked list
iv.	Remove an element from the end of the doubly linked list